+++ plugins/provisioners/ansible/cap/guest/suse/ansible_install.rb
@@ -6,7 +6,7 @@ module VagrantPlugins
         module SUSE
           module AnsibleInstall
 
-            def self.ansible_install(machine, install_mode, ansible_version)
+            def self.ansible_install(machine, install_mode, ansible_version, pip_args)
               if install_mode != :default
                 raise Ansible::Errors::AnsiblePipInstallIsNotSupported
               else