+++ test/unit/plugins/guests/windows/cap/rsync_test.rb
@@ -19,7 +19,7 @@ describe "VagrantPlugins::GuestWindows::Cap::RSync" do
 
   describe ".rsync_pre" do
     it 'makes the guestpath directory with mkdir' do
-      communicator.expect_command("mkdir '/sync_dir'")
+      communicator.expect_command("mkdir -p '/sync_dir'")
       described_class.rsync_pre(machine, guestpath: '/sync_dir')
     end
   end